feat: Add rent token initialization to vault
This commit adds a new function `initialize_rent_token` to the `nft_vault` module. This function initializes the rent token for the vault by setting the `rent_token_mint` field in the vault account. It takes the `rent_token_mint` account as an argument and assigns it to the vault's `rent_token_mint` field.

The purpose of this change is to enable the vault to track and manage rent payments using a specific token. This functionality is necessary for the proper functioning of the NFT vault.